requirements.txt
google-cloud-bigquery==3.13.0